blog/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
sqlparse==0.4.4
